Pytorch/requirements.txt
Cython                           0.29.16                       
h5py                             2.9.0                         
imageio                          2.5.0                                
matplotlib                       3.1.0               
mmcv                             0.2.16              
numpy                            1.16.4              
opencv-python                    4.1.1.26            
pandas                           0.25.0              
Pillow                           6.0.0                                 
scikit-image                     0.15.0              
scikit-learn                     0.23.1              
scipy                            1.5.0                               
tensorboardX                     1.6                            
torch                            1.1.0               
torchvision                      0.4.2                          
tqdm                             4.32.1                    
